Add common module, move apply_masks, get_chunker_ids, slice_codepoints to common.utils
