+++ src/orchestrator/streaming.rs
@@ -26,7 +26,7 @@ use tokio::sync::{broadcast, mpsc};
 use tokio_stream::wrappers::{BroadcastStream, ReceiverStream};
 use tracing::{debug, error, info, instrument, warn, Instrument, Span};
 
-use super::{get_chunker_ids, Context, Error, Orchestrator, StreamingClassificationWithGenTask};
+use super::{Context, Error, Orchestrator, StreamingClassificationWithGenTask};
 use crate::{
     clients::{
         chunker::{tokenize_whole_doc_stream, ChunkerClient, DEFAULT_CHUNKER_ID},
@@ -38,7 +38,10 @@ use crate::{
         DetectorParams, GuardrailsTextGenerationParameters, TextGenTokenClassificationResults,
         TokenClassificationResult, UNSUITABLE_INPUT_MESSAGE,
     },
-    orchestrator::unary::{input_detection_task, tokenize},
+    orchestrator::{
+        common::get_chunker_ids,
+        unary::{input_detection_task, tokenize},
+    },
     pb::{caikit::runtime::chunkers, caikit_data_model::nlp::ChunkerTokenizationStreamResult},
 };
