+++ src/orchestrator/unary.rs
@@ -25,9 +25,9 @@ use futures::{
 use tracing::{debug, error, info, instrument, Instrument, Span};
 
 use super::{
-    apply_masks, get_chunker_ids, ChatDetectionTask, Chunk, ClassificationWithGenTask, Context,
-    ContextDocsDetectionTask, DetectionOnGenerationTask, Error, GenerationWithDetectionTask,
-    Orchestrator, TextContentDetectionTask,
+    ChatDetectionTask, Chunk, ClassificationWithGenTask, Context, ContextDocsDetectionTask,
+    DetectionOnGenerationTask, Error, GenerationWithDetectionTask, Orchestrator,
+    TextContentDetectionTask,
 };
 use crate::{
     clients::{
@@ -48,6 +48,7 @@ use crate::{
         TextContentDetectionResult, TextGenTokenClassificationResults, TokenClassificationResult,
         UNSUITABLE_INPUT_MESSAGE,
     },
+    orchestrator::common::{apply_masks, get_chunker_ids},
     pb::caikit::runtime::chunkers,
 };
